124 Autherization - Gates

+++ app/Http/Controllers/PostController.php
@@ -37,9 +37,9 @@ class PostController extends Controller
      */
     public function create()
     {
+        $this->authorize('create_post');
+
         $categories = Category::all();
-
-
         return view('create', compact('categories'));
     }
 
@@ -48,6 +48,8 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create_post');
+
         $request->validate([
             'image' => ['required', 'max:2028', 'image'],
             'title' => ['required', 'max:255'],
@@ -87,6 +89,8 @@ class PostController extends Controller
      */
     public function edit(string $id)
     {
+        $this->authorize('edit_post');
+
         $post = Post::find($id);
         $categories = Category::all();
 
@@ -98,7 +102,7 @@ class PostController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        // return $request;
+        $this->authorize('edit_post');
 
         $request->validate([
             'title' => ['required', 'max:255'],
@@ -135,6 +139,8 @@ class PostController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('delete_post');
+
         $post = Post::findOrFail($id);
         $post->delete();
 
@@ -143,12 +149,16 @@ class PostController extends Controller
 
     function trashed()
     {
+        $this->authorize('delete_post');
+
         $posts = Post::onlyTrashed()->get();
         return view('trashed', compact('posts'));
     }
 
     public function restore($id)
     {
+        $this->authorize('delete_post');
+
         $post = Post::onlyTrashed()->findOrFail($id);
 
         $post->restore();
@@ -158,6 +168,8 @@ class PostController extends Controller
 
     public function forceDelete($id)
     {
+        $this->authorize('delete_post');
+        
         $post = Post::onlyTrashed()->findOrFail($id);
         File::delete(public_path($post->image));
         $post->forceDelete();
